# openapi.yaml
openapi: 3.0.3
info:
  title: Battle Cards Game API
  description: |
    A Rock Paper Scissors battle card game implemented using microservices architecture.
    All services are accessed through the Nginx API Gateway at http://localhost:8080.
    
    **Authentication**: Most endpoints require JWT Bearer tokens. Register or login to receive a token.
  version: 1.0.0
  contact:
    name: Advanced Software Engineering Project
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api
    description: Local development server (via Nginx Gateway)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/register or /api/auth/login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
          description: User ID
        username:
          type: string
          description: Username
        created_at:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - username
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        access_token:
          type: string
          description: JWT bearer token
        token_type:
          type: string
          enum: [bearer]
          description: Token type (OAuth2-style)
        expires_in:
          type: integer
          description: Token expiration time in seconds
        user:
          $ref: '#/components/schemas/User'
      required:
        - message
        - access_token
        - token_type
        - expires_in
        - user
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register new user
      description: Create a new user account and receive JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                  description: Username (3-50 characters)
                password:
                  type: string
                  minLength: 4
                  maxLength: 128
                  description: Password (4-128 characters)
              required:
                - username
                - password
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Auth
      summary: User login
      description: Authenticate user and receive JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/profile:
    get:
      tags:
        - Auth
      summary: Get user profile
      description: Retrieve current user's profile information
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Auth
      summary: Update user profile
      description: Update user profile (currently supports password update)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  minLength: 4
                  maxLength: 128
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/validate:
    post:
      tags:
        - Auth
      summary: Validate JWT token
      description: Validate a JWT token (internal service use)
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  username:
                    type: string
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Card Service
...